refactor: Move `TorrentActor` to its own file & create wrapper struct
